move StellarSigner and eliminate trait
