+++ cmd/soroban-cli/src/config/secret.rs
@@ -4,14 +4,10 @@ use std::{io::Write, str::FromStr};
 use stellar_strkey::ed25519::{PrivateKey, PublicKey};
 
 use crate::print::Print;
-use crate::utils::transaction_hash;
-use crate::xdr::{self, DecoratedSignature};
 use crate::{
-    signer::{self, LocalKey},
+    signer::{self, LocalKey, SignerKind, StellarSigner},
     utils,
 };
-
-use super::network::Network;
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
@@ -164,32 +160,6 @@ impl Secret {
 
     pub fn test_seed_phrase() -> Result<Self, Error> {
         Self::from_seed(Some("0000000000000000"))
-    }
-}
-
-pub struct StellarSigner {
-    kind: SignerKind,
-    printer: Print,
-}
-
-pub enum SignerKind {
-    Local(LocalKey),
-}
-
-#[async_trait::async_trait]
-impl signer::SignTx for StellarSigner {
-    async fn sign_tx(
-        &self,
-        txn: &xdr::Transaction,
-        network: &Network,
-    ) -> Result<DecoratedSignature, signer::types::Error> {
-        let tx_hash = transaction_hash(txn, &network.network_passphrase)?;
-        let hex_hash = hex::encode(tx_hash);
-        self.printer
-            .infoln(format!("Signing transaction with hash: {hex_hash}"));
-        match &self.kind {
-            SignerKind::Local(key) => key.sign_tx_hash(tx_hash),
-        }
     }
 }
